Removes image field and image delete btn on team edit if image doesn't exist

+++ team/index.php
@@ -14,9 +14,11 @@ $connection->close();
             <form id="add-team-form" method="POST" action="../edit.php">
                 <?php require('../feedback.php') ?>
                 <div class="form-head">
-                    <div class="team-logo">
-                        <img src="/<?= $teamItem['teamLogo'];?>" alt="">
-                    </div>
+                    <?php if (!empty($teamItem['teamLogo'])) : ?>
+                        <div class="team-logo">
+                            <img src="/<?= $teamItem['teamLogo'];?>" alt="">
+                        </div>
+                    <?php endif; ?>
                     <h1>Edit Your Team</h1>
                 </div>
                 <div class="form-body">
@@ -27,7 +29,9 @@ $connection->close();
                     <div class="form-field">
                         <label>Team Logo</label>
                         <input type="file" name="teamLogo">
-                        <button><a href="/delete-image.php?id=<?= $teamID; ?>">Delete Image</a></button>
+                        <?php if (!empty($teamItem['teamLogo'])) : ?>
+                            <button><a href="/delete-image.php?id=<?= $teamID; ?>">Delete Image</a></button>
+                        <?php endif; ?>
                     </div>
                     <div class="form-field">
                         <label>Team Name</label>
